+++ plugin/core/gvrender_core_pic.c
@@ -343,7 +343,7 @@ static void pic_ellipse(GVJ_t * job, pointf * A, int filled)
 		PS2INCH(A[0].y));
 }
 
-static void pic_bezier(GVJ_t *job, pointf *A, int n, int filled) {
+static void pic_bezier(GVJ_t *job, pointf *A, size_t n, int filled) {
     (void)filled;
 
     pointf V[4];
@@ -355,9 +355,9 @@ static void pic_bezier(GVJ_t *job, pointf *A, int n, int filled) {
     PF2P(A[0], p);
     gvprintf(job, "move to (%d, %d)", p.x, p.y);
     /* write subsequent points */
-    for (int i = 0; i + 3 < n; i += 3) {
+    for (size_t i = 0; i + 3 < n; i += 3) {
         V[0] = V[3];
-        for (int j = 1; j <= 3; j++) {
+        for (size_t j = 1; j <= 3; j++) {
             V[j].x = A[i + j].x;
             V[j].y = A[i + j].y;
         }
